requirements.txt
torch==2.0.1
torchvision==0.15.2
torchmetrics==1.0.2
pytorch-lightning==2.0.6
Pillow==9.3.0
pandas==2.0.3
matplotlib==3.7.3
numpy==1.25.2
scikit-learn==1.3.0
einops==0.8.0
gdown==5.2.0
pycocotools==2.0.6
opencv-python==4.9.0.80
git+https://github.com/gautamchitnis/cocoapi.git@cocodataset-master#subdirectory=PythonAPI&egg=pycocotools